// src/app/componentes/models/Tarjeta.ts
export class Tarjeta {
    id: number;
    titulo: string;
    lugarCargue: string;
    lugarDestino: string;
    peso: string;
    imagen: string;
//     constructor(id:number, titulo:string,descripcion:string,imagen:string){
//         this.id=id;
//         this.titulo=titulo;
//         this.descripcion=descripcion;
//         this.imagen=imagen
//     }
//
}